# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tinder API",
        "description": "API Documentation for Tinder App",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/liked": {
            "get": {
                "tags": [
                    "Liked"
                ],
                "summary": "Get list of liked people",
                "operationId": "0c833622c79852486b816e9cda769124",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of liked people"
                    }
                }
            }
        },
        "/api/people/{id}/like": {
            "post": {
                "tags": [
                    "Reaction"
                ],
                "summary": "Like a person",
                "operationId": "fa11d9179c5e8070276ace0e13030e74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Liked successfully"
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
        "/api/people/{id}/dislike": {
            "post": {
                "tags": [
                    "Reaction"
                ],
                "summary": "Dislike a person",
                "operationId": "25b6dbd30caca054e2b1c81daa2334e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disliked successfully"
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
        "/api/people": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get list of recommended people",
                "operationId": "75e2bf0e39ef85d0c4de2c6a888f9fa6",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude for location filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude for location filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius_km",
                        "in": "query",
                        "description": "Radius in km for location filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of people",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Get a single person",
                "operationId": "071d181674956e8cd9193e7327782262",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Liked",
            "description": "API endpoints for liked people"
        },
        {
            "name": "Reaction",
            "description": "API endpoints for liking/disliking people"
        },
        {
            "name": "People",
            "description": "API endpoints for People"
        }
    ]
}